refactor(shape): answer review, first pass
Five of the twelve comments, the ones that stand alone. The arithmetic
relocation the review is mostly about is a separate change.

**The nursery group is gone from `[workspace.lints.clippy]`**, and with it
every allow that existed only to silence it: thirteen files lose a
`redundant_pub_crate` or `missing_const_for_fn` attribute, including the
crate-wide one on `corvid_shape/src/lib.rs`. The group and
`unreachable_pub` want opposite things of a private module's items, so
each of those allows was a rule arguing with another rule rather than
with the code.

**`inflate` is `radius * slope.hypot(1)`**, which is what it always
meant. `corvid_fixed::hypot` computes it from the exact sum of squares,
so the Q32 widening this spelled out by hand happens where the scale is
known.

**`slope_of` uses `tan`.** It was computing a tangent from `sin_cos` and
a division, including a hand-written guard for the quarter turn where
the cosine is zero -- and `Pitch32::tan` already saturates there. The
guard is the type's answer now rather than this function's.

**`Aabb::centre` is `self.min + self.half_extent()`.** It works, and my
comment claiming it would saturate was wrong: the centre lies between
the corners, both of which are points already, so only the *extent*
needs widening and that is `half_extent`'s business.

**The signed-normalized conversion takes an `i8`.** It is the narrowest
integer that spells the three values the type holds, so the width says
what the conversion is for; `direction(0, 1, 0)` still compiles, since
what an unsuffixed literal needs is one candidate impl rather than a
particular one.

`I24F8::to_i16f16` is new, in a `convert` module beside the scalars --
the tangent comes back as a Q8 and a frustum works in Q16. Its own
module rather than `point/mod.rs` because a conversion that changes the
binary scale can clamp, which is a different kind of operation from the
arithmetic around it, and there will be more of them.

+++ crates/corvid_fixed/src/fixed/signed/macros.rs
@@ -334,14 +334,15 @@ macro_rules! define_signed {
         /// What it buys is the bare-number spelling at a call site, so
         /// `direction(0, 1, 0)` reads as the axis it is.
         ///
-        /// `i32` and no other width, for the reason the point builders take
-        /// one integer type each: an unsuffixed literal reaches an
-        /// `impl Into<Self>` parameter as an inference variable and rustc
-        /// commits it only when exactly one candidate applies. A second impl
-        /// would make `direction(0, 1, 0)` stop compiling.
-        impl From<i32> for $name {
+        /// `i8` and no other width. It is the narrowest integer that spells
+        /// the three values this type actually holds, which is the honest
+        /// width for a conversion whose whole range is `-1 ..= 1` -- and one
+        /// impl rather than several is what lets an unsuffixed literal reach
+        /// an `impl Into<Self>` parameter at all, since rustc commits an
+        /// inference variable only when exactly one candidate applies.
+        impl From<i8> for $name {
             #[inline]
-            fn from(value: i32) -> Self {
+            fn from(value: i8) -> Self {
                 if value >= 1 {
                     Self::MAX
                 } else if value <= -1 {
